requirements.txt
python-magic>=0.4.27
pyclamd>=0.4.0          
cryptography>=41.0.0
